【新增】增加PCR检测统计信息及图表展示。PCR精度及间隔图示没有找到合适的设置X轴标签的方法，暂时以数字代替。

+++ src/main/java/m2tk/assistant/dbi/mapper/PCREntityMapper.java
@@ -32,7 +32,7 @@ public class PCREntityMapper implements RowMapper<PCREntity>
         entity.setId(rs.getLong("id"));
         entity.setPid(rs.getInt("pid"));
         entity.setPosition(rs.getLong("pct"));
-        entity.setValue(rs.getInt("value"));
+        entity.setValue(rs.getLong("value"));
         return entity;
     }
 }
